refact(helpers): refact helpers
