+++ src/Service/Base/Jssdk/Client.php
@@ -15,7 +15,6 @@ use EaseBaidu\Kernel\Exceptions\RuntimeException;
 use EaseBaidu\Kernel\BaseClient;
 use EaseBaidu\Kernel\Support\Str;
 use EaseBaidu\Kernel\Traits\InteractsWithCache;
-use EaseBaidu\Kernel\Support;
 
 class Client extends BaseClient
 {
@@ -140,7 +139,7 @@ class Client extends BaseClient
             return $this->url;
         }
 
-        return Support\current_url();
+        return current_url();
     }
 
     /**
